+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+## new
+* improve: Improve Jsonx's isEmptyJSON(String), isNotEmptyJSON(String), isEmptyJSONObject(String), isNotEmptyJSONObject(String), isEmptyJSONArray(String), isNotEmptyJSONArray(String), isJSONObject(String), isNotJSONObject(String), isJSONArray(String), isEmptyJSONArray( String) method performance and memory consumption, especially for long strings, the improvement is obvious
+
 ## 1.0.0-alpha01
 * new: Added JSONArray sequence support in kotlin, for example:
 ```kotlin